fix: CurrentSize/Size/FileSize int→long（防止 >2GB 文件静默截断）
根源修复：shared-spec.json v1.3.0 为所有文件大小字段添加 csharpType: long
- TypeMapper 支持 csharpType 覆盖
- 6 个实体字段：FileEntry.CurrentSize, VersionRecord.Size, SyncQueue.FileSize,
  RemoteSnapshot.Size, BackupLog.FileSize, OfflineCache.FileSize
- FileIndexService 移除 (int)size 强制转换
- ClientDbContext SyncQueueItem/RemoteSnapshot 同步修改

+++ CloudPan.Client/Models/ClientDbContext.cs
@@ -55,7 +55,7 @@ public class SyncQueueItem
     public int Operation { get; set; } // 0=Upload, 1=Download, 2=Delete
     public int Priority { get; set; }
     public int? BaseVersion { get; set; }
-    public int? FileSize { get; set; }
+    public long? FileSize { get; set; }
     public int RetryCount { get; set; }
     public string? LastError { get; set; }
     public string CreatedAt { get; set; } = DateTime.UtcNow.ToString("O");
@@ -67,7 +67,7 @@ public class RemoteSnapshot
     public string Path { get; set; } = "";
     public int Type { get; set; }
     public string? Hash { get; set; }
-    public int Size { get; set; }
+    public long Size { get; set; }
     public int Version { get; set; }
     public int State { get; set; }
 }
